+++ duckql/functions/to_date.py
@@ -9,19 +9,18 @@ from ..properties.constant import Constant
 from ..structures.cast_operator import CastOperator
 
 
-class DateFormat(BaseFunction):
-    obj: Literal['functions.DateFormat'] = 'functions.DateFormat'
+class ToDate(BaseFunction):
+    obj: Literal['functions.ToDate'] = 'functions.ToDate'
     property: Union[Property, Constant, BaseFunction, CastOperator]
     format: str
     alias: str = None
 
     @validator('format', pre=True)
     def sanitize_format(cls, v):
-        return cls.escape(v, ['%'])
+        return cls.escape(v)
 
     def to_sql(self) -> str:
-        # TODO: postgresql notation for to_char
-        sql = f"DATE_FORMAT({self.property}, '{self.format}')"
+        sql = f"to_date({self.property}, '{self.format}')"
 
         if self.alias is not None:
             sql = f"{sql} AS {self.alias}"